fix  bug delete student

+++ controllers/deletestudent.php
@@ -49,12 +49,16 @@ if (!isset($_SESSION['loggedin'])) {
     include 'connectdb.php';
     $mahocsinh = $_POST['mahocsinh'];
 
-    // Check age of student [15,20]
 
+    // delete PHIEUDIEM before delete Student 
+    $sql = "delete from PHIEUDIEM
+    where mahocsinh = '$mahocsinh'";
+    if($conn->query($sql) === TRUE)
+    {
     $sql = " delete from HOCSINH
     where mahocsinh = '$mahocsinh';
     ";
-    
+    }
     if ($conn->query($sql) === TRUE) {
       echo '<script type="text/JavaScript">
               document.getElementById("div-inform").innerHTML += "Xóa Học Viên Thành Công";
